% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbSendMultiUpdate.R
\name{dbSendMultiUpdate}
\alias{dbSendMultiUpdate}
\title{Run multiple SQL query actions within the DB environment.}
\usage{
dbSendMultiUpdate(conn, conn_type = "JDBC", sql_path)
}
\arguments{
\item{conn}{A database connection object}

\item{conn_type}{A string indicating if the connection type is a JDBC or ODBC connection - accepts "JDBC" or "ODBC" as arguments, defaults to "JDBC"}

\item{sql_path}{Character vector pointing to SQL script}
}
\value{
No object returned - for production automation
}
\description{
Requires a DBI workflow that creates a connection with a relational database per DBI: https://dbi.r-dbi.org/. This function is an extension of the RJDBC::dbSendUpdate function.
}
