% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbGetMultiQuery.R
\name{dbGetMultiQuery}
\alias{dbGetMultiQuery}
\title{Query a database with multiple query actions and one, final, select statement.}
\usage{
dbGetMultiQuery(connection, sql_file_path, pattern = NULL, replacement = NULL)
}
\arguments{
\item{connection}{A database connection object}

\item{sql_file_path}{Character vector pointing to SQL script}

\item{pattern}{A character object you would like to substitute in the SQL script - this is not required and defaults to NULL. Calls gsub under the hood, so use regex for pattern identification}

\item{replacement}{A character vector replacing the pattern specified - this is not required and defaults to NULL. Calls gsub under the hood, so use regex for pattern identification}
}
\value{
A data object
}
\description{
Requires a DBI workflow that creates a connection with a relational database per DBI: https://dbi.r-dbi.org/. This function is an extension of the DBI::dbGetQuery and RJDBC::dbSendUpdate functions.
}
\examples{
dbGetMultiQuery(conn, "~/path_to/file.sql")
dbGetMultiQuery(conn, "~/path_to/file.sql", pattern = "state = [A-Z](2)", replacement = "state = MD")
}
