% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbSendMultiUpdate.R
\name{dbSendMultiUpdate}
\alias{dbSendMultiUpdate}
\title{Run multiple SQL query actions within the DB environment.}
\usage{
dbSendMultiUpdate(connection, sql_file_path)
}
\arguments{
\item{connection}{A database connection object}

\item{sql_file_path}{Character vector pointing to SQL script}
}
\value{
No object returned - for production automation
}
\description{
Requires a DBI workflow that creates a connection with a relational database per DBI: https://dbi.r-dbi.org/. This function is an extension of the RJDBC::dbSendUpdate function.
}
